restructured code which belongs in Agent
