+++ src/out.json
@@ -154,7 +154,7 @@
     {
       "track_id": 0,
       "category_id": 0,
-      "track_len": 33,
+      "track_len": 35,
       "steps": [
         0,
         1,
@@ -188,7 +188,9 @@
         29,
         30,
         31,
-        32
+        32,
+        33,
+        34
       ]
     }
   ],
@@ -245,8 +247,8 @@
       "image_id": "frame_16",
       "category_id": 0,
       "bbox": [
-        31.22771722686995,
-        100.12492197250387,
+        31.227717226869963,
+        100.12492197250396,
         1,
         1
       ],
@@ -314,8 +316,8 @@
       "image_id": "frame_19",
       "category_id": 0,
       "bbox": [
-        32.744075514610394,
-        107.70329614269,
+        32.744075514610444,
+        107.70329614269005,
         1,
         1
       ],
@@ -406,8 +408,8 @@
       "image_id": "frame_23",
       "category_id": 0,
       "bbox": [
-        39.69897016243856,
-        141.4213562373096,
+        39.69897016243855,
+        141.42135623730954,
         1,
         1
       ],
@@ -475,8 +477,8 @@
       "image_id": "frame_26",
       "category_id": 0,
       "bbox": [
-        61.93892527675272,
-        122.06555615733708,
+        61.93892527675262,
+        122.06555615733704,
         1,
         1
       ],
@@ -544,7 +546,7 @@
       "image_id": "frame_29",
       "category_id": 0,
       "bbox": [
-        67.25592448538978,
+        67.25592448538966,
         103.07764064044159,
         1,
         1
@@ -613,8 +615,8 @@
       "image_id": "frame_32",
       "category_id": 0,
       "bbox": [
-        68.77228277312994,
-        101.98039027185571,
+        68.77228277312989,
+        101.98039027185568,
         1,
         1
       ],
@@ -682,8 +684,8 @@
       "image_id": "frame_35",
       "category_id": 0,
       "bbox": [
-        64.35292528604134,
-        119.2686044187657,
+        64.35292528604143,
+        119.26860441876562,
         1,
         1
       ],
@@ -797,8 +799,8 @@
       "image_id": "frame_40",
       "category_id": 0,
       "bbox": [
-        56.9380676935905,
-        172.0465053408526,
+        56.938067693590384,
+        172.04650534085255,
         1,
         1
       ],
@@ -820,8 +822,8 @@
       "image_id": "frame_41",
       "category_id": 0,
       "bbox": [
-        53.1602850203427,
-        204.0176127617325,
+        55.114576874409636,
+        221.0626759717437,
         1,
         1
       ],
@@ -843,8 +845,8 @@
       "image_id": "frame_42",
       "category_id": 0,
       "bbox": [
-        45.18628140278768,
-        177.556410786635,
+        44.67155940042146,
+        160.41325016025343,
         1,
         1
       ],
@@ -866,8 +868,8 @@
       "image_id": "frame_43",
       "category_id": 0,
       "bbox": [
-        44.15065044514747,
-        210.64212019201136,
+        45.041418909878814,
+        207.9617460082693,
         1,
         1
       ],
@@ -889,8 +891,8 @@
       "image_id": "frame_44",
       "category_id": 0,
       "bbox": [
-        34.16959142786998,
-        183.7966512066052,
+        30.887711898143955,
+        152.4157017866303,
         1,
         1
       ],
@@ -912,8 +914,8 @@
       "image_id": "frame_45",
       "category_id": 0,
       "bbox": [
-        35.71058882041471,
-        217.79902829020213,
+        34.31392258689343,
+        198.90538665286056,
         1,
         1
       ],
@@ -935,8 +937,54 @@
       "image_id": "frame_46",
       "category_id": 0,
       "bbox": [
-        23.847205696969258,
-        190.59704540687974,
+        68.24841922367025,
+        147.6007223577372,
+        1,
+        1
+      ],
+      "area": 1,
+      "segmentation": [],
+      "iscrowd": 0,
+      "track_id": 0,
+      "trackmap_index": 0,
+      "vid_id": 0,
+      "track_color": [
+        255,
+        255,
+        255
+      ],
+      "displaced": true
+    },
+    {
+      "id": 33,
+      "image_id": "frame_47",
+      "category_id": 0,
+      "bbox": [
+        64.77249974377845,
+        161.99450908174535,
+        1,
+        1
+      ],
+      "area": 1,
+      "segmentation": [],
+      "iscrowd": 0,
+      "track_id": 0,
+      "trackmap_index": 0,
+      "vid_id": 0,
+      "track_color": [
+        255,
+        255,
+        255
+      ],
+      "displaced": false
+    },
+    {
+      "id": 34,
+      "image_id": "frame_48",
+      "category_id": 0,
+      "bbox": [
+        61.86564305133053,
+        176.4892877903516,
         1,
         1
       ],
